+++ src/main/java/com/example/preproj/controller/RequestRestController.java
@@ -1,18 +1,12 @@
 package com.example.preproj.controller;
 
-import com.example.preproj.model.Request;
 import com.example.preproj.model.User;
-import com.example.preproj.service.request.RequestService;
-import com.example.preproj.service.user.UserServiceImp;
+import com.example.preproj.service.EmailService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -20,23 +14,33 @@ import java.util.List;
 public class RequestRestController {
 
     @Autowired
-    private RequestService requestService;
-
-    @Autowired
-    private UserServiceImp userService;
+    private EmailService emailService;
 
     @GetMapping("/")
     public ResponseEntity<List<User>> list() {
-        List<Request> list = requestService.reqUsers();
-        List<User> userList = new ArrayList<>();
-        for (Request request : list) {
-            userList.add(userService.show(request.getUserId()));
-        }
-        return new ResponseEntity<>(userList, HttpStatus.OK);
+        return new ResponseEntity<>(emailService.allRequest(), HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Boolean> button(@PathVariable("id") Integer id) {
-        return new ResponseEntity<>(requestService.show(id) != null, HttpStatus.OK);
+        return new ResponseEntity<>(emailService.show(id), HttpStatus.OK);
+    }
+
+    @GetMapping("/mail/{id}")
+    public ResponseEntity<String> mail(@PathVariable("id") Integer id) {
+        emailService.request(id);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<String> update(@PathVariable("id") Integer id) {
+        emailService.acceptedApplication(id);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> delete(@PathVariable("id") Integer id) {
+        emailService.rejectedApplication(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
