+++ src/main/java/com/example/preproj/service/RequestService.java
@@ -1,10 +1,9 @@
-package com.example.preproj.service.request;
+package com.example.preproj.service;
 
 import com.example.preproj.model.Request;
 import com.example.preproj.model.Role;
 import com.example.preproj.model.User;
 import com.example.preproj.repo.RequestRepo;
-import com.example.preproj.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +13,7 @@ import java.util.Optional;
 import java.util.Set;
 
 @Service
-public class RequestServiceImp implements RequestService {
+public class RequestService {
 
     @Autowired
     UserService userService;
@@ -22,17 +21,17 @@ public class RequestServiceImp implements RequestService {
     @Autowired
     RequestRepo requestRepo;
 
-    @Override
+
     public void add(int userId) {
         requestRepo.save(new Request(userId));
     }
 
-    @Override
+
     public List<Request> reqUsers() {
         return requestRepo.findAll();
     }
 
-    @Override
+
     public void update(int userId) {
         Set<Role> set = new HashSet<>();
         User user = userService.show(userId);
@@ -44,12 +43,12 @@ public class RequestServiceImp implements RequestService {
 
     }
 
-    @Override
+
     public void delete(int userId) {
         requestRepo.deleteById(userId);
     }
 
-    @Override
+
     public Request show(int id) {
         Optional<Request> optionalUser = requestRepo.findById(id);
         return optionalUser.get();
