+++ src/main/java/com/example/preproj/service/UserService.java
@@ -1,4 +1,4 @@
-package com.example.preproj.service.user;
+package com.example.preproj.service;
 
 import com.example.preproj.model.Role;
 import com.example.preproj.model.User;
@@ -13,7 +13,7 @@ import java.util.Optional;
 import java.util.Set;
 
 @Service
-public class UserServiceImp implements UserService {
+public class UserService {
 
     @Autowired
     private PasswordEncoder bCryptPasswordEncoder;
@@ -21,17 +21,17 @@ public class UserServiceImp implements UserService {
     @Autowired
     private UserRepo userRepo;
 
-    @Override
+
     public void delete(int id) {
         userRepo.deleteById(id);
     }
 
-    @Override
+
     public List<User> allUsers() {
        return userRepo.findAll();
     }
 
-    @Override
+
     public boolean save(User user, String[] role) {
         User userFromDB = userRepo.findByUserName(user.getUsername());
         Set<Role> set = new HashSet<>();
@@ -49,13 +49,13 @@ public class UserServiceImp implements UserService {
         return true;
     }
 
-    @Override
+
     public boolean updateSave(User user) {
         userRepo.save(user);
         return true;
     }
 
-    @Override
+
     public User show(int id) {
         Optional<User> optionalUser = userRepo.findById(id);
         return optionalUser.get();
