--- src/main/java/com/example/preproj/service/email/EmailService.java
@@ -1,5 +1,0 @@
-package com.example.preproj.service.email;
-
-public interface EmailService {
-    void sendSimpleEmail(final String toAddress, final String subject, final String message);
-}
