--- src/main/java/com/example/preproj/service/request/RequestService.java
@@ -1,18 +1,0 @@
-package com.example.preproj.service.request;
-
-import com.example.preproj.model.Request;
-
-import java.util.List;
-
-
-public interface RequestService {
-    void add(int userId);
-
-    List<Request> reqUsers();
-
-    void update(int userId);
-
-    void delete(int userId);
-
-    public Request show(int id);
-}
